refactor: split infer config regressions
